Properly cull duplicates
hiit.py was cleverly attempting to avoid duplicates by maintaining a
`seen_ex` set to store seen exercises... but not populating it during
the loop.

This commit gets rid of that set and instead just uses the `workout`
list.

This ends up being slightly less efficient processor-wise and slightly
more efficient memory-wise; but the effect is negligible in both cases
as long as the data set is so small. We'll revisit once you're
generating workouts with thousands of exercises :)